requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.7
aiosqlite==0.19.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
httpx==0.25.2
python-multipart==0.0.6